requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
dj-rest-auth==6.0.0
Django==5.0.7
django-allauth==64.0.0
django-cors-headers==4.4.0
djangorestframework==3.15.2
dnspython==2.6.1
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
jwt==1.3.1
marshmallow==3.21.3
mongoengine==0.28.2
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
pycparser==2.22
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
